==> 01-twitter-archive-cleaning.ipynb <==
# From archives of twitter api, extract tweets in polish, with no attached media

import pandas as pd
import pathlib
import bz2
import json
import concurrent.futures
from tqdm import tqdm
import re
import langdetect
import pickle
import fastprogress

source_dir = pathlib.Path('/mnt/dane/twitter/')

source_files = [x for x in source_dir.iterdir() if '.tar.pkl' in x.name]

def get_texts(sfile):
    tweets = [tweet for sub in pickle.load(open(sfile, 'rb')) for tweet in sub]
    texts = []
    for t in tweets:
        retw = t.get('retweeted_status')
        current = retw or t
        if current.get('extended_tweet'):
            continue
        if current['truncated']:
            continue
        entities = current.get('entities', {})
        if entities.get('media') or len(entities.get('urls', []))>0:
            continue
        texts.append(current['text'])
    return texts


sub_texts = [get_texts(f) for f in tqdm(source_files)]

texts = [t for sub in sub_texts for t in sub]

def get_proba_of_lang(t: str, lang:str = 'pl'):
    try:
        langs = langdetect.detect_langs(t)
    except langdetect.lang_detect_exception.LangDetectException:
        return 0
    return next(iter([x.prob for x in langs if x.lang == lang]), 0)


with concurrent.futures.ProcessPoolExecutor(16) as pool:
    results = list(tqdm(pool.map(get_proba_of_lang, texts), total=len(texts)))

df = pd.DataFrame({'text': texts, 'probability': results})

df = df.loc[df.probability.gt(0.99)]

filtered = df.text

len(filtered)

at_mention = re.compile('@[a-zA-Z0-9_]*')
def remove_handles(t):
    return re.sub(at_mention, '@anonymized_account', t)

def replace_newlines(t):
    return t.replace('\n', '\\n')

URL = re.compile(' ?https?://.*')
def replace_urls(t):
    return re.sub(URL, ' @url', t)

processed = filtered.apply(remove_handles)

processed = processed.apply(replace_urls)

# processed = processed.apply(replace_newlines)

labels = pd.Series(index=processed.index).fillna('?')

resdf = pd.concat([labels, processed], axis=1)

# # texts = [t.replace('\\n', '\n') for t in texts]
# texts = [t.replace('\\"', '"') for t in texts]
# URL = re.compile(' ?https?://.*')
# texts = [URL.sub(' @url', t) for t in texts]

len(resdf)

resdf = resdf.drop_duplicates()

len(resdf)

resdf.to_csv('/home/tomasz/nlp/poleval/csv/unsup-twitter.csv', index=False, header=False)

labelled = pd.read_csv('/home/tomasz/nlp/poleval/csv/task1_train.csv', header=None, names=resdf.columns)

alldf = pd.concat([resdf, labelled], axis=0)

alldf.head()

alldf.tail()

alldf.to_csv('/home/tomasz/nlp/poleval/csv/unsup.csv', index=False, header=False)
