==> 03.2-experiment-results.ipynb <==
# Browse results of experiments

import pymongo
import pandas as pd
pd.options.display.max_colwidth=100

db = pymongo.MongoClient('mongodb://localhost')['ulmfit_experiments']

results = list(db['completed'].find({'type': 'Poleval1'}))
flattened = [dict(id=x['_id'], **x['results'], **x['params']) for x in results]
df = pd.DataFrame(flattened).set_index('id')

df = df.assign(last_phase_cycles = df.training_phases.apply(lambda x: x[-1]['cyc_len']))
df.drop_mult.fillna(0.5, inplace=True)
df.cv_fold_num.fillna(0, inplace=True)
df.bidir.fillna(False, inplace=True)
df.bs.fillna(40, inplace=True)
df.max_len.fillna(70*20, inplace=True)
df = df.assign(num_phases = df.phase_stats.apply(lambda x: len(x)))
# df.rnn_output_layers.fillna((-1,), inplace=True)
to_drop = ['time_submitted', 'time_completed', 'dataset_path', 'encoder_subdir', 'train_losses', 'phase_stats', 'training_phases']
sdf = df.loc[:, ~df.columns.isin(to_drop)].copy()
phases = lambda num: pd.DataFrame([x for phase in df.loc[num].phase_stats for x in phase])

sdf = sdf.assign(best_avg_precision = df.phase_stats.apply(lambda x: max([ep['avg_precision'] for stage in x for ep in stage])))

sdf.sort_values('best_avg_precision', ascending=False)
