==> 04-run-best-config.ipynb <==
# Best config params were chosen, now running it 5 times with different subsets of train set as validation set

from fastai.text import * 

from ulmfit_experiments import experiments
from ulmfit_experiments import sentencepiece_tok as spm
from ulmfit_experiments.runner import MongoExperimentRunner
from sklearn.metrics import *

torch.manual_seed(0)
torch.backends.cudnn.deterministic = True
torch.backends.cudnn.benchmark = False
np.random.seed(0)

runner = MongoExperimentRunner(experiments.ExperimentCls.factory, '/home/tomasz/nlp/trained_models/')

db = runner._get_db()

source_id = 137

source_config = db[runner.completed_collection].find_one(source_id)

c = {'type': deepcopy(source_config['type']), 'params': deepcopy(source_config['params'])}

c

# ex = experiments.ExperimentCls.factory(c['type'], c['params'])

# results, learn = ex.run()

for fold in range(1, 5): 
    print('fold ' +  str(fold))
    c = {'type': deepcopy(source_config['type']), 'params': deepcopy(source_config['params'])}
    c['params']['cv_fold_num'] = fold
    ex = experiments.ExperimentCls.factory(c['type'], c['params'])
    results, learn = ex.run()
    learn.save('poleval-submission-'+str(fold))
    
    preds = learn.get_preds(DatasetType.Valid, ordered=True)
    dfs = ex.get_dfs(fold)
    target = dfs[1][ex.label_col]
    predicted_proba = preds[0][:, 1].numpy()
    fscores = pd.Series({t: f1_score(target, predicted_proba > t) for t in np.arange(0.01,0.99, 0.01)})
    fscores.to_csv(f'/home/tomasz/nlp/poleval/submission/fold{fold}-valid-fscores.csv')
    
    preds = learn.get_preds(DatasetType.Test, ordered=True)
    predicted_proba = preds[0][:, 1].numpy()
    pd.Series(predicted_proba).to_csv(f'/home/tomasz/nlp/poleval/submission/fold{fold}-probas.csv')
