==> 05-ensemble.ipynb <==
# For each CV fold:
- reading probabilities, 
- establishing the best threshold (to maximize f1-score)
- making a decision

## then taking mode of decisions

import pandas as pd
import os
import glob

probas_files = glob.glob('/home/tomasz/nlp/poleval/submission/fold*probas.csv')

fscores_files = glob.glob('/home/tomasz/nlp/poleval/submission/fold*-valid-fscores.csv')

probas = pd.concat({os.path.basename(f): pd.read_csv(f, header=None, usecols=[1], squeeze=True) for f in probas_files}, axis=1)

probas.columns = probas.columns.str.replace('-probas.csv', '')

fscores = pd.concat({os.path.basename(f): pd.read_csv(f, header=None, index_col=0) for f in fscores_files}, axis=1)

thresholds = fscores.idxmax()

thresholds.index = pd.Index([f'fold{num}' for num in range(5)])

(probas > thresholds).astype(int).mode(axis=1).to_csv('/home/tomasz/nlp/poleval/submission/task01/results.txt', index=False, header=False)

# probas.mean(axis=1).gt(thresholds.mean()).astype(int).to_csv('/home/tomasz/nlp/poleval/submission/task01/mean_thr.txt', index=False, header=False)
